Move block controls logic to a separate class
